requirements.txt
pulumi>=3.133.0,<4
pulumi-datarobot>=0.2.5,<0.3

datarobot>=3.5.2,<3.6
datarobot-predict>=1.9.2,<1.10
datarobotx-idp>=0.2.11,<0.3
datarobot-drum>=1.12.2,<1.13

pydantic>=2.9.2,<2.10
pydantic-settings>=2.5.2,<2.6
eval_type_backport>=0.2.0,<0.3

openai>=1.31.2,<2
# Constrained by datarobot-drum
pandas>=2.0.3,<3

streamlit>=1.39.0,<2
plotly>=5.24.1,<6
tiktoken>=0.8.0,<0.9
scikit-learn>=1.5.2,<2
textblob>=0.18.0,<0.19

jupyterlab>=4.2.5,<5
papermill>=2.6.0,<3

pytest==8.0.1
pytest-mock==3.14.0
mypy==1.11.2
types-PyYAML==6.0.12
types-requests==2.32.0.20241016
pandas-stubs>=2.2.3,<3
ruff==0.6.9
